+++ che3le/__init__.py
@@ -1,0 +1,28 @@
+'''
+Atrificial Neural Network (ANN) package
+
+This package is composed of the following subpackages and modules:  
+
+* `tensor` module
+* `utils` subpackage
+    * `functions` module
+    * `processing` module
+    * `simulation` module
+    * `visualization` module (not implemented yet)
+    * `evaluation` module (not implemented yet)
+    * `validation` module (not implemented yet)
+* `extensions` subpackage  
+    * datasets module 
+    * dataloader module
+    * `transforms` module 
+* `nn` subpackage
+    * `module` module
+    * `activation` module
+    * `loss` module
+    * `optimizer` module
+    * `linear` module  
+    * `parameter` module
+'''
+
+from che3le.tensor import int32, float32, float64, int64, uint8
+from che3le.utils.functions import *